Issue #235: add checks that free users cannot handle apikeys

+++ project/apps/identity_provider/views/token_generator.py
@@ -9,41 +9,54 @@ class ApiKeyManager(permissions.IsAuthenticated, views.APIView):
 
     def post(self, request):
         user = request.user
-        token, created = ApiKey.objects.get_or_create(
-            user=user, last_modified=datetime.utcnow()
-        )
-        return JsonResponse(
-            {
+        if self._is_a_free_user(user):
+            data={}
+            status=403
+        else:
+            token, created = ApiKey.objects.get_or_create(
+                user=user, last_modified=datetime.utcnow()
+            )
+            data = {
                 "token": token.key,
                 "created": created,
                 "last_modified": token.last_modified,
-            },
-            status=200,
-        )
+            }
+            status=200
+        return JsonResponse(data, status=status)
 
     def put(self, request):
         user = request.user
-        token = ApiKey.objects.filter(user=user)
-        if token:
-            token.delete()
-        token, created = ApiKey.objects.get_or_create(
-            user=user, last_modified=datetime.utcnow()
-        )
-        return JsonResponse(
-            {
+        if self._is_a_free_user(user):
+            data={}
+            status=403
+        else:
+            token = ApiKey.objects.filter(user=user)
+            if token:
+                token.delete()
+            token, created = ApiKey.objects.get_or_create(
+                user=user, last_modified=datetime.utcnow()
+            )
+            data = {
                 "token": token.key,
                 "created": created,
                 "last_modified": token.last_modified,
-            },
-            status=200,
-        )
+            }
+            status=200
+        return JsonResponse(data, status=status)
+
 
     def delete(self, request):
         user = request.user
-        token = ApiKey.objects.filter(user=user)
-        if token:
-            token.delete()
-            status = 200
+        if self._is_a_free_user(user):
+            status=403
         else:
-            status = 500
+            token = ApiKey.objects.filter(user=user)
+            if token:
+                token.delete()
+                status = 200
+            else:
+                status = 500
         return JsonResponse(data={}, status=status)
+
+    def _is_a_free_user(self, user):
+        return 'free' in [g.name.lower() for g in user.group_set.all()]
